requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.7
Flask==1.1.2
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.16.0
MarkupSafe==1.1.1
pprint==0.1
pydash==4.7.6
python-decouple==3.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
